Remove AddThis widget
The AddThis widget was bringing in lots of cookies, whihc we would have had to made toggleable and added into a new functionality section of the cookie banner. Looking at it, it looks a bit phishy, and we've already made the decision to remove it off of other sites previously.